renaming of files and classes (to align with terms used in demo paper)
