+++ mallm/decision_protocols/ranked_voting.py
@@ -1,9 +1,8 @@
 import logging
-from collections import Counter
 from typing import Any, Optional
 
 from mallm.agents.panelist import Panelist
-from mallm.decision_protocol.protocol import DecisionProtocol
+from mallm.decision_protocols.protocol import DecisionProtocol
 from mallm.models.discussion.ResponseGenerator import ResponseGenerator
 from mallm.utils.config import Config
 from mallm.utils.enums import DecisionAlteration
@@ -12,12 +11,12 @@ from mallm.utils.types import Agreement, VotingResult, VotingResultList, WorkerF
 logger = logging.getLogger("mallm")
 
 
-class ApprovalVoting(DecisionProtocol):
+class RankedVoting(DecisionProtocol):
     """
-    The Approval Voting decision protocol allows panelists to approve any number of solutions after a certain number of turns.
+    The Ranked Voting decision protocol allows panelists to rank their preferences among a set of answers after a certain number of turns.
     """
 
-    _name = "approval_voting"
+    _name = "ranked_voting"
 
     def __init__(
         self,
@@ -47,6 +46,7 @@ class ApprovalVoting(DecisionProtocol):
         final_answers_with_confidence, voting_process_string = (
             self.generate_final_answers(agreements, question, task)
         )
+
         decision, final_answer, results, voting_process_string = (
             self.vote_with_alterations(
                 final_answers_with_confidence,
@@ -54,7 +54,7 @@ class ApprovalVoting(DecisionProtocol):
                 task,
                 voting_process_string,
                 self._name,
-                ResponseGenerator.generate_approval_voting_prompt,
+                ResponseGenerator.generate_ranking_prompt,
                 config.voting_protocols_with_alterations,
             )
         )
@@ -71,33 +71,34 @@ class ApprovalVoting(DecisionProtocol):
         all_votes: dict[str, VotingResult],
         alteration: DecisionAlteration,
         final_answers: list[str],
-        votes: list[int],
+        votes: list[list[int]],
     ) -> dict[str, VotingResult]:
-        winners = []
-        most_approved = -1
-        if votes:
-            # Count approvals for each answer
-            # Get the most common approval count
-            approval_counts = Counter(votes)
-            most_common = approval_counts.most_common()
-            most_approved = most_common[0][0]
-            most_approved_count = most_common[0][1]
+        # Calculate the score for each answer based on the rankings
+        scores = [0] * len(final_answers)
+        for ranking_list in votes:
+            for rank, idx in enumerate(ranking_list):
+                scores[idx] += (
+                    min(5, self.total_agents) - rank
+                )  # Score 5 for the 1st rank, 4 for the 2nd, etc.
 
-            # Check if there are multiple winners with the same vote count
-            winners = [
-                candidate
-                for candidate, count in approval_counts.items()
-                if count == most_approved_count
-            ]
-        if len(winners) == 1:
+        # Find the answer with the highest score
+        highest_score = max(scores)
+        index = scores.index(highest_score)
+        best_answers = [
+            final_answers[i] for i, score in enumerate(scores) if score == highest_score
+        ]
+
+        # If there's a tie, pick the first answer among the best
+        # If all panelists agree on the best answer finished else go for another round
+        if len(best_answers) == 1:
             all_votes[alteration.value] = VotingResult(
                 votes=votes,
-                most_voted=most_approved,
-                final_answer=final_answers[most_approved],
+                most_voted=index,
+                final_answer=final_answers[index],
                 agreed=True,
             )
             logger.info(
-                f"Most approved answer from agent {self.panelists[most_approved].short_id}"
+                f"Selected answer from agent {self.panelists[index].short_id} with {highest_score} points"
             )
         else:
             all_votes[alteration.value] = VotingResult(
@@ -118,16 +119,16 @@ class ApprovalVoting(DecisionProtocol):
         voting_process_string: str,
     ) -> tuple[str, Any, bool, str]:
         success = False
-        approval_list = [
-            int(a.strip())
-            for a in vote_str.split(",")
-            if 0 <= int(a.strip()) < len(final_answers)
-        ]
-        if approval_list:
-            vote.extend(approval_list)
+        # Split the ranking and convert to a list of integers
+        ranking_list = list(map(int, vote_str.strip().split()))
+        if (
+            all(0 <= rank < len(final_answers) for rank in ranking_list)
+            and len(ranking_list) <= 5
+        ):
+            vote.append(ranking_list)
             logger.info(
-                f"{panelist.persona} approved answers from {[self.panelists[a].persona for a in approval_list]}"
+                f"{panelist.persona} ranked answers: {[self.panelists[a].persona for a in ranking_list]}"
             )
-            voting_process_string += f"{panelist.persona} approved answers from {[self.panelists[a].persona for a in approval_list]}\n"
+            voting_process_string += f"{panelist.persona} ranked answers: {[self.panelists[a].persona for a in ranking_list]}\n"
             success = True
         return vote_str, vote, success, voting_process_string
